+++ R/geouy.R
@@ -10,4 +10,6 @@
 NULL
 
 ## quiets concerns of R CMD check re: the .'s that appear in pipelines
-if (getRversion() >= "2.15.1") utils::globalVariables(c(".", "label", "link", "localidad"))
+if (getRversion() >= "2.15.1") utils::globalVariables(
+  c(".", "label", "link", "localidad", "lat", "lon")
+  )