+++ NEWS.md
@@ -5,6 +5,8 @@
 ## geouy v0.2.6 (in progress)
 
 * Add Municipios 2010 and 2015 to metadata 
+* Update geocode_ide_uy() because server of IDEuy change
+* Add a reverse_ide_uy() function for reverse geocoding 
 
 ## geouy v0.2.5 (2021-08-12)